Refactored reindexer into a module (fixes #262)
